rename files and add comments
