+++ src/components/FileProcessor.js
@@ -1,6 +1,6 @@
 import { LitElement, html, css } from "lit";
 import * as procFuncs2 from "./ProcFuncs.js"; // array of funcs exported from ProcFuncs.js
-import * as procFuncs1 from "../util/json-to-d3-hierarchy.js"; // array of funcs exported from json-to-d3-hierarchy.js
+import * as procFuncs1 from "../util/persons-to-d3-hierarchy.js"; // array of funcs exported from persons-to-d3-hierarchy.js
 // const procFuncs is a merge of procFuncs1 + procFuncs2
 const procFuncs = { ...procFuncs1, ...procFuncs2 };
