+++ src/util/persons-to-d3-hierarchy.js
